Split out GeobodyFeature.py
