+++ resqpy/organize/__init__.py
@@ -1,3 +1,4 @@
 __all__ = ['OrganizationFeature',]
 
 from .OrganizationFeature import OrganizationFeature
+from .GeobodyFeature import GeobodyFeature
